Fire loadSettings() on options save
Now that the extension options are loaded at browser startup the options need to be reloaded if they are changed. It's as simple as getting the background page and launching the function. All the functions in background.js go with what ever the global variables are. LoadSettings simply makes the variables the new options.